fix: Migrate all components to TypeScript
