+++ lib/componentDocs/generateComponentDocs.js
@@ -24,11 +24,10 @@ const byDisplayName = (acc, doc) => ({
 
 (async () => {
   const componentDocs = await globsToReactDocgen({
-    include: 'lib/components/**/*.{js,ts,tsx}',
+    include: 'lib/components/**/*.{ts,tsx}',
     exclude: [
       'lib/components/private/**/*',
-      'lib/components/index.js',
-      '**/*.{css,docs,test}.js',
+      '**/*.{docs,test}.{ts,tsx}',
       '**/*.css.js.d.ts'
     ]
   });
