+++ lib/components/Box/Box.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import classnames from 'classnames';
-import withTheme from '../private/withTheme';
-import Reset, { Props as ResetProps } from '../Reset/Reset';
+import ThemeConsumer from '../ThemeConsumer/ThemeConsumer';
+import Reset, { ResetProps } from '../Reset/Reset';
 import styles from './Box.css.js';
 import {
   HorizontalSpacingVariants,
@@ -12,7 +12,7 @@ import {
   BorderColorVariants
 } from '../../themes/theme';
 
-interface Props extends ResetProps {
+export interface BoxProps extends ResetProps {
   paddingTop?: SpacingVariants;
   paddingBottom?: SpacingVariants;
   paddingLeft?: HorizontalSpacingVariants;
@@ -25,53 +25,53 @@ interface Props extends ResetProps {
   borderRadius?: BorderRadiusVariants;
   backgroundColor?: BackgroundColorVariants;
   borderColor?: BorderColorVariants;
-  className?: string;
 }
 
-export default withTheme(
-  class Box extends Component<Props> {
-    static displayName = 'Box';
+export default class Box extends Component<BoxProps> {
+  static displayName = 'Box';
 
-    render() {
-      const {
-        theme,
-        paddingTop,
-        paddingBottom,
-        paddingLeft,
-        paddingRight,
-        marginTop,
-        marginBottom,
-        marginLeft,
-        marginRight,
-        borderWidth,
-        borderRadius,
-        backgroundColor,
-        borderColor = 'standard',
-        className = '',
-        ...restProps
-      } = this.props;
+  render() {
+    const {
+      paddingTop,
+      paddingBottom,
+      paddingLeft,
+      paddingRight,
+      marginTop,
+      marginBottom,
+      marginLeft,
+      marginRight,
+      borderWidth,
+      borderRadius,
+      backgroundColor,
+      borderColor = 'standard',
+      className,
+      ...restProps
+    } = this.props;
 
-      return (
-        <Reset
-          className={classnames(
-            className,
-            styles.root,
-            theme.atoms.backgroundColor[backgroundColor!],
-            theme.atoms.borderColor[borderColor!],
-            theme.atoms.borderWidth[borderWidth!],
-            theme.atoms.borderRadius[borderRadius!],
-            theme.atoms.marginTop[marginTop!],
-            theme.atoms.marginRight[marginRight!],
-            theme.atoms.marginBottom[marginBottom!],
-            theme.atoms.marginLeft[marginLeft!],
-            theme.atoms.paddingTop[paddingTop!],
-            theme.atoms.paddingRight[paddingRight!],
-            theme.atoms.paddingBottom[paddingBottom!],
-            theme.atoms.paddingLeft[paddingLeft!]
-          )}
-          {...restProps}
-        />
-      );
-    }
+    return (
+      <ThemeConsumer>
+        {theme => (
+          <Reset
+            className={classnames(
+              className,
+              styles.root,
+              theme.atoms.backgroundColor[backgroundColor!],
+              theme.atoms.borderColor[borderColor!],
+              theme.atoms.borderWidth[borderWidth!],
+              theme.atoms.borderRadius[borderRadius!],
+              theme.atoms.marginTop[marginTop!],
+              theme.atoms.marginRight[marginRight!],
+              theme.atoms.marginBottom[marginBottom!],
+              theme.atoms.marginLeft[marginLeft!],
+              theme.atoms.paddingTop[paddingTop!],
+              theme.atoms.paddingRight[paddingRight!],
+              theme.atoms.paddingBottom[paddingBottom!],
+              theme.atoms.paddingLeft[paddingLeft!]
+            )}
+            {...restProps}
+          />
+        )}
+      </ThemeConsumer>
+    );
   }
-);
+}
