+++ lib/components/ThemeTokensConsumer/ThemeTokensConsumer.tsx
@@ -1,25 +1,21 @@
 import React, { Component, ReactChild } from 'react';
-import ThemeContext from '../private/ThemeContext';
+import ThemeConsumer from '../ThemeConsumer/ThemeConsumer';
 import { Tokens } from '../../themes/theme';
 
-interface Props {
+export interface ThemeTokensConsumerProps {
   children(tokens: Tokens): ReactChild;
 }
 
-export default class ThemeTokensConsumer extends Component<Props> {
+export default class ThemeTokensConsumer extends Component<
+  ThemeTokensConsumerProps
+> {
   static displayName = 'ThemeTokensConsumer';
 
   render() {
     return (
-      <ThemeContext.Consumer>
-        {theme => {
-          if (theme === null) {
-            throw new Error('No theme passed');
-          }
-
-          return this.props.children(theme.tokens);
-        }}
-      </ThemeContext.Consumer>
+      <ThemeConsumer>
+        {theme => this.props.children(theme.tokens)}
+      </ThemeConsumer>
     );
   }
 }
