+++ lib/components/icons/ChevronIcon/ChevronIcon.tsx
@@ -1,14 +1,15 @@
 import React, { Component } from 'react';
 import classnames from 'classnames';
+import { Omit } from 'utility-types';
 import Icon, { IconProps } from '../Icon/Icon';
 import ChevronSvg from './ChevronSvg';
 import styles from './ChevronIcon.css.js';
 
-interface Props extends IconProps {
+export interface ChevronIconProps extends Omit<IconProps, 'svgComponent'> {
   direction?: 'up' | 'down' | 'left' | 'right';
 }
 
-export default class ChevronIcon extends Component<Props> {
+export default class ChevronIcon extends Component<ChevronIconProps> {
   static displayName = 'ChevronIcon';
 
   render() {
