+++ lib/components/icons/ChevronIcon/ChevronSvg.tsx
@@ -1,6 +1,8 @@
-import React, { Component } from 'react';
+import React, { Component, AllHTMLAttributes } from 'react';
 
-export default class ChevronSvg extends Component {
+export type ChevronSvgProps = AllHTMLAttributes<SVGElement>;
+
+export default class ChevronSvg extends Component<ChevronSvgProps> {
   static displayName = 'ChevronSvg';
 
   render() {
