+++ lib/components/icons/ErrorIcon/ErrorIcon.tsx
@@ -1,8 +1,11 @@
 import React, { Component } from 'react';
+import { Omit } from 'utility-types';
 import Icon, { IconProps } from '../Icon/Icon';
 import ErrorSvg from './ErrorSvg';
 
-export default class ErrorIcon extends Component<IconProps> {
+export type ErrorIconProps = Omit<IconProps, 'svgComponent'>;
+
+export default class ErrorIcon extends Component<ErrorIconProps> {
   static displayName = 'ErrorIcon';
 
   render() {
