+++ lib/components/icons/Icon/Icon.tsx
@@ -1,18 +1,15 @@
 import React, { Component, ComponentType } from 'react';
+import { Omit } from 'utility-types';
 import classnames from 'classnames';
-import Box from '../../Box/Box';
-import withTheme, { WithThemeProps } from '../../private/withTheme';
+import Box, { BoxProps } from '../../Box/Box';
+import ThemeConsumer from '../../ThemeConsumer/ThemeConsumer';
 import styles from './Icon.css.js';
 import { TextSize, FillVariants, SizeVariants } from '../../../themes/theme';
 
-export interface IconProps {
+export interface IconProps extends Omit<BoxProps, 'size'> {
   size?: TextSize | 'fill';
   inline?: boolean;
   fill?: FillVariants;
-  className?: string;
-}
-
-interface Props extends WithThemeProps, IconProps {
   svgComponent: ComponentType;
 }
 
@@ -21,44 +18,47 @@ const isSizeVariant = (
   sizeAtom: string
 ): sizeAtom is SizeVariants => Object.keys(atom).indexOf(sizeAtom) > -1;
 
-export default withTheme(
-  class Icon extends Component<Props> {
-    static displayName = 'Icon';
+export default class Icon extends Component<IconProps> {
+  static displayName = 'Icon';
 
-    render() {
-      const {
-        className,
-        theme,
-        size,
-        svgComponent,
-        inline = false,
-        fill,
-        ...restProps
-      } = this.props;
-      const sizeAtom = `${size || 'standard'}Text${inline ? 'Inline' : ''}`;
-      const widthAtom = isSizeVariant(theme.atoms.width, sizeAtom)
-        ? theme.atoms.width[sizeAtom]
-        : '';
-      const heightAtom = isSizeVariant(theme.atoms.height, sizeAtom)
-        ? theme.atoms.height[sizeAtom]
-        : '';
-
-      return (
-        <Box
-          component={svgComponent}
-          className={classnames(
+  render() {
+    return (
+      <ThemeConsumer>
+        {theme => {
+          const {
             className,
-            widthAtom,
-            heightAtom,
-            theme.atoms.fill[fill!],
-            {
-              [styles.fillSize]: size === 'fill',
-              [styles.inline]: inline
-            }
-          )}
-          {...restProps}
-        />
-      );
-    }
+            size,
+            svgComponent,
+            inline = false,
+            fill,
+            ...restProps
+          } = this.props;
+          const sizeAtom = `${size || 'standard'}Text${inline ? 'Inline' : ''}`;
+          const widthAtom = isSizeVariant(theme.atoms.width, sizeAtom)
+            ? theme.atoms.width[sizeAtom]
+            : '';
+          const heightAtom = isSizeVariant(theme.atoms.height, sizeAtom)
+            ? theme.atoms.height[sizeAtom]
+            : '';
+
+          return (
+            <Box
+              component={svgComponent}
+              className={classnames(
+                className,
+                widthAtom,
+                heightAtom,
+                theme.atoms.fill[fill!],
+                {
+                  [styles.fillSize]: size === 'fill',
+                  [styles.inline]: inline
+                }
+              )}
+              {...restProps}
+            />
+          );
+        }}
+      </ThemeConsumer>
+    );
   }
-);
+}
