+++ lib/components/icons/InfoIcon/InfoIcon.tsx
@@ -1,8 +1,11 @@
 import React, { Component } from 'react';
+import { Omit } from 'utility-types';
 import Icon, { IconProps } from '../Icon/Icon';
 import InfoSvg from './InfoSvg';
 
-export default class InfoIcon extends Component<IconProps> {
+export type InfoIconProps = Omit<IconProps, 'svgComponent'>;
+
+export default class InfoIcon extends Component<InfoIconProps> {
   static displayName = 'InfoIcon';
 
   render() {
