+++ lib/components/icons/InfoIcon/InfoSvg.tsx
@@ -1,6 +1,8 @@
-import React, { Component } from 'react';
+import React, { Component, AllHTMLAttributes } from 'react';
 
-export default class InfoSvg extends Component {
+export type ChevronSvgProps = AllHTMLAttributes<SVGElement>;
+
+export default class InfoSvg extends Component<ChevronSvgProps> {
   static displayName = 'InfoSvg';
 
   render() {
