+++ lib/components/icons/TickCircleIcon/TickCircleIcon.tsx
@@ -1,8 +1,11 @@
 import React, { Component } from 'react';
+import { Omit } from 'utility-types';
 import Icon, { IconProps } from '../Icon/Icon';
 import TickCircleSvg from './TickCircleSvg';
 
-export default class TickCircleIcon extends Component<IconProps> {
+export type TickCircleIconProps = Omit<IconProps, 'svgComponent'>;
+
+export default class TickCircleIcon extends Component<TickCircleIconProps> {
   static displayName = 'TickCircleIcon';
 
   render() {
