+++ lib/components/private/ThemeContext.ts
@@ -1,7 +1,4 @@
 import { createContext } from 'react';
 import { Theme } from './../../themes/theme';
 
-const context = createContext<Theme | null>(null);
-context.displayName = 'Theme';
-
-export default context;
+export default createContext<Theme | null>(null);
