--- lib/themes/utils/toUnit.js
@@ -1,2 +1,0 @@
-export const px = str =>
-  str.length === 0 || /^0$|[a-z]+/.test(str) ? str : `${str}px`;
