+++ lib/themes/utils/toUnit.ts
@@ -1,0 +1,1 @@
+export const px = (num: number) => `${num}${num !== 0 ? 'px' : ''}`;
